Split context into container and context
