+++ autowire/__init__.py
@@ -1,9 +1,11 @@
+from .container import Container
 from .context import Context
 from .exc import ResourceNotProvidedError
 from .resource import Resource
 
 __all__ = [
     "Context",
+    "Container",
     "Resource",
     "ResourceNotProvidedError",
 ]
